requirements.txt
python-chess
easydict:
